+++ src/main/java/dao/UsuarioDAOimpl.java
@@ -6,6 +6,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
+import java.util.ArrayList;
 import java.util.List;
 
 public class UsuarioDAOimpl implements UsuarioDAO {
@@ -28,30 +29,60 @@ public class UsuarioDAOimpl implements UsuarioDAO {
 
     @Override
     public void eliminarUsuario(int id) throws Exception {
-        String sql = "delete from Usuario where id = ?";
-        try (Connection con = ConnectionManager.getConnection();
-             PreparedStatement ps = con.prepareStatement(sql)){
-            ps.setInt(1, id);
-            ps.executeUpdate();
-            System.out.println("Usuario eliminado: " + id);
+        PrestamoDAO prestamoDAO = new PrestamoDAOimpl();
+        if(prestamoDAO.comprobarPrestamosUsuarios(id)){
+            System.out.println("El usuario tiene prestamos activos");
+        }else {
+            String sql = "delete from Usuario where id = ?";
+            try (Connection con = ConnectionManager.getConnection();
+                 PreparedStatement ps = con.prepareStatement(sql)) {
+                ps.setInt(1, id);
+                ps.executeUpdate();
+                System.out.println("Usuario eliminado: " + id);
 
+            }
         }
     }
 
     @Override
     public List<Usuario> listarUsuario() throws Exception {
-        return List.of();
+        List<Usuario> usuarios = new ArrayList<>();
+        String sql = "select * from Usuario";
+        try (Connection con = ConnectionManager.getConnection();
+             PreparedStatement ps = con.prepareStatement(sql)) {
+            ResultSet rs = ps.executeQuery();
+            while (rs.next()) {
+                usuarios.add(new Usuario(rs.getInt("id"), rs.getString("nombre")));
+            }
+        }
+        return usuarios;
+    }
+
+    @Override
+    public Usuario getUsuarioById(int id) throws Exception {
+        String sql = "select nombre from Usuario where id = ?";
+        try(Connection con = ConnectionManager.getConnection();
+            PreparedStatement ps = con.prepareStatement(sql)){
+            ps.setInt(1, id);
+            try(ResultSet rs = ps.executeQuery()){
+                if (rs.next()) {
+                    return new Usuario(rs.getInt("id"),
+                            rs.getString("nombre"));
+                }
+            }
+        }
+        return null;
     }
 
     @Override
     public void actualizarUsuario(Usuario usuario) throws Exception {
         String sql = "update Usuario set nombre = ? where id = ?";
         try (Connection con = ConnectionManager.getConnection();
-             PreparedStatement ps = con.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)){
+             PreparedStatement ps = con.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
             ps.setString(1, usuario.getNombre());
             ps.setInt(2, usuario.getId());
             ps.executeUpdate();
-            try (ResultSet rs = ps.getGeneratedKeys()){
+            try (ResultSet rs = ps.getGeneratedKeys()) {
                 if (rs.next()) usuario.setNombre(rs.getString(1));
             }
             System.out.println("Usuario actualizado: " + usuario);
